+++ SpriteClipCreator.cpp
@@ -1,13 +1,13 @@
 #include "SpriteClipCreator.h"
 
-SpriteClipCreator::SpriteClipCreator(int imgHeight, int imgWidth, 
-                                     int tileHeight, int tileWidth, 
+SpriteClipCreator::SpriteClipCreator(uint16_t imgHeight, uint16_t imgWidth, 
+                                     uint16_t tileHeight, uint16_t tileWidth, 
                                      std::vector<SDL_Rect> &spriteClip) {
-    int files = imgHeight / tileHeight;
-    int columns = imgWidth / tileWidth;
+    uint16_t files = imgHeight / tileHeight;
+    uint16_t columns = imgWidth / tileWidth;
 
-    for (int i = 0; i < files; i++) {
-        for (int j = 0; j < columns; j++) {
+    for (uint16_t i = 0; i < files; i++) {
+        for (uint16_t j = 0; j < columns; j++) {
             spriteClip.emplace_back(SDL_Rect{j*tileWidth, i*tileHeight, 
                                              tileWidth, tileHeight});
         }
